+++ src/modules/fun/index.ts
@@ -1,10 +1,13 @@
 /** @format */
 
+export type { AlwaysHasBeenOptions, AlwaysHasBeenSlot } from "./alwaysHasBeen";
+export { alwaysHasBeen } from "./alwaysHasBeen";
 export { distractedBoyfriend } from "./distractedBoyfriend";
 export { drake } from "./drake";
 export { level } from "./level";
 export { Music } from "./music";
 export { Quote } from "./quote";
+export { twoButtons } from "./twoButtons";
 export type { WelcomeCardMetaItem } from "./welcomeCard";
 export {
 	WelcomeCardBuilder,